3788 - Re-implement DeviceClassifier in vanilla JS
